added optional parameter to collections
