+++ Manatee.Trello/BoardCollection.cs
@@ -26,10 +26,11 @@ namespace Manatee.Trello
 		/// Creates a new board.
 		/// </summary>
 		/// <param name="name">The name of the board to create.</param>
-		/// <param name="source">A board to use as a template.</param>
+		/// <param name="description">(Optional) A description for the board.</param>
+		/// <param name="source">(Optional) A board to use as a template.</param>
 		/// <param name="ct">(Optional) A cancellation token for async processing.</param>
 		/// <returns>The <see cref="IBoard"/> generated by Trello.</returns>
-		public async Task<IBoard> Add(string name, IBoard source = null, CancellationToken ct = default(CancellationToken))
+		public async Task<IBoard> Add(string name, string description = null, IBoard source = null, CancellationToken ct = default(CancellationToken))
 		{
 			var error = NotNullOrWhiteSpaceRule.Instance.Validate(null, name);
 			if (error != null)
@@ -37,6 +38,7 @@ namespace Manatee.Trello
 
 			var json = TrelloConfiguration.JsonFactory.Create<IJsonBoard>();
 			json.Name = name;
+			json.Desc = description;
 			if (source != null)
 				json.BoardSource = ((Board) source).Json;
 			if (_addRequestType == EntityRequestType.Organization_Write_CreateBoard)
