+++ Manatee.Trello/CardCollection.cs
@@ -30,10 +30,8 @@ namespace Manatee.Trello
 		/// <param name="labels">(optional) A collection of labels to add to the card.</param>
 		/// <param name="ct">(Optional) A cancellation token for async processing.</param>
 		/// <returns>The <see cref="ICard"/> generated by Trello.</returns>
-		public async Task<ICard> Add(string name, string description = null, Position position = null,
-		                             DateTime? dueDate = null,
-		                             bool? isComplete = null, IEnumerable<IMember> members = null,
-		                             IEnumerable<ILabel> labels = null,
+		public async Task<ICard> Add(string name, string description = null, Position position = null, DateTime? dueDate = null,
+		                             bool? isComplete = null, IEnumerable<IMember> members = null, IEnumerable<ILabel> labels = null,
 		                             CancellationToken ct = default(CancellationToken))
 		{
 			var error = NotNullOrWhiteSpaceRule.Instance.Validate(null, name);
@@ -43,7 +41,7 @@ namespace Manatee.Trello
 			var json = TrelloConfiguration.JsonFactory.Create<IJsonCard>();
 			json.Name = name;
 			json.Desc = description;
-			json.Pos = position == null ? null : Position.GetJson((Position) position);
+			json.Pos = position == null ? null : Position.GetJson(position);
 			json.Due = dueDate;
 			json.DueComplete = isComplete;
 			json.IdMembers = members?.Select(m => m.Id).Join(",");
