+++ Manatee.Trello/Contracts/IBoardCollection.cs
@@ -12,9 +12,10 @@ namespace Manatee.Trello
 		/// Creates a new board.
 		/// </summary>
 		/// <param name="name">The name of the board to create.</param>
-		/// <param name="source">A board to use as a template.</param>
+		/// <param name="description">(Optional) A description for the board.</param>
+		/// <param name="source">(Optional) A board to use as a template.</param>
 		/// <param name="ct">(Optional) A cancellation token for async processing.</param>
 		/// <returns>The <see cref="IBoard"/> generated by Trello.</returns>
-		Task<IBoard> Add(string name, IBoard source = null, CancellationToken ct = default(CancellationToken));
+		Task<IBoard> Add(string name, string description = null, IBoard source = null, CancellationToken ct = default(CancellationToken));
 	}
 }
