+++ Manatee.Trello/Contracts/IListCollection.cs
@@ -12,8 +12,9 @@ namespace Manatee.Trello
 		/// Creates a new list.
 		/// </summary>
 		/// <param name="name">The name of the list to add.</param>
+		/// <param name="position">(Optional) The position of the list.</param>
 		/// <param name="ct">(Optional) A cancellation token for async processing.</param>
 		/// <returns>The <see cref="IList"/> generated by Trello.</returns>
-		Task<IList> Add(string name, CancellationToken ct = default(CancellationToken));
+		Task<IList> Add(string name, Position position = null, CancellationToken ct = default(CancellationToken));
 	}
 }
