+++ Manatee.Trello/Contracts/IOrganizationCollection.cs
@@ -12,9 +12,11 @@ namespace Manatee.Trello
 		/// Creates a new organization.
 		/// </summary>
 		/// <param name="displayName">The display name of the organization to add.</param>
+		/// <param name="description">(Optional) A description of the organization.</param>
+		/// <param name="name">(Optional) An org-name for the organization; like a member's username.  If omitted, Trello will generate it from <paramref name="displayName"/>.</param>
 		/// <param name="ct">(Optional) A cancellation token for async processing.</param>
 		/// <returns>The <see cref="IOrganization"/> generated by Trello.</returns>
 		/// <remarks>The organization's name will be derived from the display name and can be changed later.</remarks>
-		Task<IOrganization> Add(string displayName, CancellationToken ct = default(CancellationToken));
+		Task<IOrganization> Add(string displayName, string description = null, string name = null, CancellationToken ct = default(CancellationToken));
 	}
 }
