+++ Manatee.Trello/ListCollection.cs
@@ -22,9 +22,10 @@ namespace Manatee.Trello
 		/// Creates a new list.
 		/// </summary>
 		/// <param name="name">The name of the list to add.</param>
+		/// <param name="position">(Optional) The position of the list.</param>
 		/// <param name="ct">(Optional) A cancellation token for async processing.</param>
 		/// <returns>The <see cref="IList"/> generated by Trello.</returns>
-		public async Task<IList> Add(string name, CancellationToken ct = default(CancellationToken))
+		public async Task<IList> Add(string name, Position position = null, CancellationToken ct = default(CancellationToken))
 		{
 			var error = NotNullOrWhiteSpaceRule.Instance.Validate(null, name);
 			if (error != null)
@@ -34,6 +35,7 @@ namespace Manatee.Trello
 			json.Name = name;
 			json.Board = TrelloConfiguration.JsonFactory.Create<IJsonBoard>();
 			json.Board.Id = OwnerId;
+			json.Pos = Position.GetJson(position);
 
 			var endpoint = EndpointFactory.Build(EntityRequestType.Board_Write_AddList);
 			var newData = await JsonRepository.Execute(Auth, endpoint, json, ct);
